Remove hardcoded featured governor from homepage
He was hardcoded because there was no summary for him with the
featured field set to true. So his url was hardcoded with the
old version of CSV person urls which used the id instead of
the slug.

This change removes that so that the route code takes care
of sending the right governor there