bin/deploy: Detect if the Sinatra server startup fails
If the deploy script tried to start the Sinatra server, but it
immediately failed (e.g. because the bundle wasn't installed correctly)
the deploy script used to just get stuck at "Waiting for Sinatra to
start..."  This change fixes that - instead it will break from the loop
that waits for the server to start if it detects that the server process
has died.