Add a contact page with a contact form
The contact form uses the free service Formspree

    https://formspree.io/

All the input fields of the contact form that have a name attribute will
be emailed.

The owner has to use it once to trigger the activation of the
form, which consists of confirming the email to which the messages
from the contact form will be sent.