Add a person factory
This person factory will be in charge of creating new person
instances.

At the moment we have a coupling with the finder classes and the
objects they create (for example, people classes are coupled to
person classes because we have an explicit call to `.new` on those
classes).

This finds us constantly changing these classes when the signature
of the created classes changes (for example, having to pass an extra
argument to the constructor of the people classes because the created
class needs it). See for example commits:

   2272b581c7dfbc4387ff6e37ad14b12bc973eb36

and

    4367744a2ec05c774a1da9043c90226669bdf6c0

there are a lot of files that change because we added a new argument
to the constructor.

This is not a reason for this classes to change (and their tests, and
the app.rb, and etc. etc.), so it may be better to create factories
to be in charge of that and have the finder classes just ask the
factory to create that object for them.

This will also simplify the signature of the finder classes, from
having a number of arguments that are only used for the creation of
those objects to just passing one: the factory.