Use the person factory in the people classes
This change uses the person factory so that the tests can be
simplified and don't have a dependency on the person signature.

Using the factory also simplifies the people class signature for
same reasons (no dependency on the person signature)

Some of the tests for the people classes are not needed anymore.
In particular, those dealing with checking that an argument passed
to the entity class was set correctly now live in the factory tests.