Prepare the EP people tests for the introduction of a factory
The introduction of a person factory will simplify the class and
its tests, so this change reorganizes the test file so that the
the purely formatting changes don't make it hard to parse the
real changes

Basically when we move to person factories in the future, some of
the tests for the people classes won't be needed anymore as they
will be moved to be part of the factory class tests.

In particular, those tests dealing with checking that an argument
passed to the entity class was set correctly will make more sense
in the factory tests.