Remove the repeated test for the current term
This tests has been bugging me for a long time. First, there is no
need for it as the same test exists already in EP::Person. Second,
to arrive where it arrives it has to go through a lot of chained
methods that expose a lot of internals for no reason. Third, there
are methods for the latest term already, asserting the date.

Hence I am removing this test.