Add a method to finder to find unpublished documents
This method allows the scraper route to send the unpublished
documents to the view so that they can be scraped and pushed to
the static repo.

Documents that are unpublished are not linked anywhere on the
site. With this change, they can be loaded as a preview without
being explicitly linked from the site.

This means that all documents, linked or unlinked, are published.