Use sinatra/config_file with some simple site specific settings
We use mySociety's standard way of adding configuration data to
our sites using a yaml file. In this commit, only simple settings
are added.

At the moment, if the requirements change, we have to change the
signature of classes. As the application keeps growing, adding
more arguments to their constructors may not be a good idea.

This site specific settings separates everything that is site
specific from the site, and also can be helpful if we pass it as
an argument to classes as well, so that we can simplify the
signature of their constructors and avoid changes from cascading
the codebase when more settings are needed due to a change in
requirements.