Add a simple hash with some simple site specific settings
The settings hash is extended with the module to convert a hash
keys to methods so that we can use it in the main app to
retrieve site specific settings as an example.

In this commit, only simple settings are added.

At the moment, if the requirements change, we have to change the
signature of classes. As the application keeps growing, adding
more arguments to their constructors may not be a good idea.

This site specific settings separates everything that is site
specific from the site, and also can be helpful if we pass it as
an argument to classes as well, so that we can simplify the
signature of their constructors and avoid changes from cascading
the codebase when more settings are needed due to a change in
requirements.