Use the site specific settings in the MapIt wrapper
This commit shows how to use the site specific settings in our main
classes. The benefits of passing any site specific setting like
base url, folder names etc. like this is that we don't have to
pollute the constructor with pointless arguments and we screen
the application from internals of the class that can be handled
inside of it.

If we change the configuration or if the class needs any other
setting, that matters only to the class. This approach helps with
that.

Since Sinatra settings are not accessible from tests (and even if
they were, we want to keep these classes independent and uncoupled
from the framework) a method was added in the test helper to read
the settings from disk using yaml