Throw an error if a slug is missing or repeated
Since slug_to_person uses the person slugs as keys, we have to
ensure that a slug exists. If it doesn't, we will throw an
error, so that this is solved upstream.

We also have to ensure that a slug is unique, otherwise, the new
coming person will overwrite the person we had already saved. If
the slug is repeated, we will throw an error, so that this is
solved upstream.