Fix Mapit to EveryPolitician area ID mapping
EveryPolitician has dropped the "area/" prefix from its area IDs [1] so
update the mappings for reps and senators to remove the "area/" prefix.

This restores the area name on the reps and senate person listing pages.

To keep the tests passing I've had to update the test fixtures, which
use real data in them and use the mapit mapping that's being updated.

[1]: https://github.com/everypolitician/everypolitician-data/commit/9a75c94fb3f01a45e5616242dec9743ba96f137f