Upgrade to Ruby 2.5.3
The current version of Ruby, 2.3.1, is "In security maintenance phase
(will EOL soon!)", so upgrade to the newest available stable version of
Ruby.

I had to update Webmock as well because the version we were on is
incompatible with Ruby >= 2.4 [1]

The change in `bin/build_summaries.rb` was prompted by a new Rubocop
warning when I switched the target version to 2.5.

[1]: https://github.com/bblimke/webmock/issues/683