Tell Rubocop to ignore builds/ directory
Every time you do a deploy a static version of the site is added to the
builds/ directory. Over time this directory becomes huge and contains
loads of files. When Rubocop runs it scans all files looking for ones
that look like Ruby files. This means that Rubocop takes ages to start
when you have a big builds/ directory. So to fix this just tell Rubocop
to ignore everything in that directory.