Make email on politician profile plain text
Our partners have told us that having this as a link is confusing
because their mail client isn't setup properly, so it doesn't work as
expected.

They originally asked us if we could add a contact form to this page,
but that's going to take a lot of time and add a lot of complexity. So
as an alternative solution we're simply making the email plain text,
which makes it easier to simply copy/paste it into your email client.