Show 10 random people of each type on homepage
Rather than showing featured people on the homepage, we instead want to
show random people. However because the site is static we can't just
show one random person on page load. So instead we show 10 random
people, then the idea is we can hide 9 of them with JavaScript. Then
those 10 people will change each time the static site is rebuild.