Randomly show one of each rep type on homepage
When the page loads, only the first representative in each group is
made visible. Then, when the DOM is ready, JavaScript randomly picks
one to show, and hides all the others.

By default, web browsers tend to load image resources specified in
`<img>` elements, even if those elements (or their parents) are hidden
by CSS. To avoid loading all 30 representative images at pageload, we
substitute the `src` attribute with a `data-src` for all but the first
representative in each group. The JavaScript then swaps the `data-src`
for `src` when the representative is unhidden, which tells the browser
to request the image file.

The end result is that only the first 3 rep images are loaded, and then
up to 3 more once the JavaScript executes. (It might be less than 3 if
the "randomly" chosen rep is in fact the first, already visible, rep!)
This should hopefully keep network use down, while also making the page
work fine for visitors without JavaScript enabled or (more likely) with
very slow loading or partial/broken JavaScript files.