Add script for converting EveryPolitician data to CSV
Uses the EveryPolitician gem to get person data for the House of
Representatives and the Senate from EveryPolitician and convert it to a
CSV format.

This will allow us to switch to getting Representative and Senate data
from CSV files, rather than EveryPolitician, which will make updates to
the data easier.